Extract namespace handling into a separate module and mix-in class.
